backend/requirements.txt
fastapi
python-dotenv
langchain-core>=0.1.52,<0.2.0
langchain-community>=0.0.38
langchain
langchain-openai
langgraph
stripe-agent-toolkit
pydantic[email]
uvicorn